+++ Engine/src/vm/virtualmachine.cpp
@@ -132,8 +132,8 @@ namespace prev {
 
 		const float minVal = 0.0f;
 		const float maxVal = 16.0f;
-		const float width = 200;
-		const float height = 150;
+		const int width = 200;
+		const int height = 150;
 		const int numVals = 128;
 
 		m_LineGraphUpdate	= new LineGraph(minVal, maxVal, Vec2i(width, height), numVals);
